Move Rubix to HDFS module
