+++ lib/trino-hdfs/src/main/java/io/trino/hdfs/rubix/RubixConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.rubix;
+package io.trino.hdfs.rubix;
 
 import com.qubole.rubix.spi.CacheConfig;
 import io.airlift.configuration.Config;
