+++ lib/trino-hdfs/src/main/java/io/trino/hdfs/rubix/RubixModule.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.rubix;
+package io.trino.hdfs.rubix;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.inject.Binder;
