+++ lib/trino-hdfs/src/test/java/io/trino/hdfs/rubix/TestRubixConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.rubix;
+package io.trino.hdfs.rubix;
 
 import com.google.common.collect.ImmutableMap;
 import com.qubole.rubix.spi.CacheConfig;
